move test error struct back into mocks_test.go
